embeddings/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astunparse==1.6.3
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
Deprecated==1.2.14
durationpy==0.9
fastapi
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.9.0
ftfy==6.3.0
gast==0.6.0
google-auth==2.35.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
kubernetes==31.0.0
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
namex==0.0.8
networkx==3.4.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai-whisper==20240930
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pillow==11.0.0
posthog==3.7.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.9.2
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.40.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
# tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf-slim==1.1.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
whisper==1.1.10
wrapt==1.16.0
zipp==3.20.2
